Move get_text function to tracing.py #817
 * this function is used only in tracing and testing
 * also sort imports and minor related cleanups
